Populate working
populate the pathology card is working
